Properly handle renamed files
The Git class now stores the list of renamed files in the GitChanged
class, and those are treated as modified, but old file name is used for
the old step, and the new file name is used for the new step. And
instead of copying that logic in a few places, its now in the GitChanged
class itself.

Bug: T202470
Change-Id: I9da7d517599972074dee2c007a0b22318b76033a