Avoid comment reference to deprecated suite.xml
The "modern" config is in phpunit.xml.dist, and suite.xml will be going
away soon.

I'm not sure if what the TODO says (it taking forever) is still true
with pcov, but the comment seems mostly harmless and trimming the
coverage filter is probably a good idea regardless (although not so
urgent), so I'm leaving it there.

Bug: T395470
Change-Id: I7a25cbde87538502515b075311beb5a5f7bfaa53