Fix "PHP Fatal error: Declaration … must be compatible" on PHP 8.3
The tool fatals on PHP 8.3 every time.

> PHP Fatal error: Declaration of
> MediaWiki\Tool\PatchCoverage\CheckCommand::execute(InputInterface $input, OutputInterface $output)
> must be compatible with
> Symfony\Component\Console\Command\Command::execute(InputInterface $input, OutputInterface $output): int
> in /src/CheckCommand.php on line 190

This is because upstream symfony/console relaeased a very un-cool
breaking change in Symfony 7.3, adding a native return type to
Command::execute() and thus fatally breaking any existing projects.

The CheckCommand class has no test coverage.

Note that even with the added test, the fatal error does not found
happen in the CI for this repo, because there is a dev dependency on
ockcyp/covers-validator 1.7, which depends on symfony/console v6.

In the mwcore-phpunit-coverage-patch Jenkins job, where this is
installed separately, it will pick the newer symfony/console v7,
and thus exposed to the breaking change.

View the fatal error at
https://integration.wikimedia.org/ci/job/mwcore-phpunit-coverage-patch/93529/console
Or reproduce locally as:
* checkout this patch, undoing the src fix but keep the test.
* remove "ockcyp/covers-validator" from composer.json
* run `composer update`
* run `composer test`, to observe fatal error
* run `git checkout .` to restore src fix
* run `composer test`, to observe success

Bug: T358666
Change-Id: I86fa083fb9314251c7f6426e1a1dd284d0958803